LPS-73675 Rename
