+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/resources/META-INF/resources/admin/templates/data-provider-parameter.js
@@ -1,6 +1,6 @@
 Liferay.Loader.require(
 	'frontend-js-metal-web/metal-soy-bundle/build/bundle',
-	'dynamic-data-lists-form-web/admin/templates/data_provider_parameter.es',
+	'dynamic-data-lists-form-web/admin/templates/data-provider-parameter.es',
 	function(MetalSoyBundle, DataProviderParameter) {
 		if (!window.DDLDataProviderParameter) {
 			window.DDLDataProviderParameter = {};
