+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/resources/META-INF/resources/admin/templates/field-options-toolbar.js
@@ -1,6 +1,6 @@
 Liferay.Loader.require(
 	'frontend-js-metal-web/metal-soy-bundle/build/bundle',
-	'dynamic-data-lists-form-web/admin/templates/field_options_toolbar.es',
+	'dynamic-data-lists-form-web/admin/templates/field-options-toolbar.es',
 	function(MetalSoyBundle, FieldOptionsToolbar) {
 		if (!window.DDLFieldSettingsToolbar) {
 			window.DDLFieldSettingsToolbar = {};
