+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/resources/META-INF/resources/admin/templates/rule-builder.es.js
@@ -1,0 +1,19 @@
+import Component from 'metal-component';
+import Soy from 'metal-soy';
+
+import templates from './rule-builder.soy';
+
+let RuleBuilderTemplates = [];
+
+for (let template in templates) {
+	if (template !== 'templates') {
+		class C extends Component {};
+		Soy.register(C, templates, template);
+		RuleBuilderTemplates.push({
+			key: template,
+			component: C
+		});
+	}
+}
+
+export default RuleBuilderTemplates;
