+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/resources/META-INF/resources/admin/templates/rule-builder.js
@@ -1,12 +1,14 @@
 Liferay.Loader.require(
 	'frontend-js-metal-web/metal-soy-bundle/build/bundle',
-	'dynamic-data-lists-form-web/admin/templates/rule_builder.es',
+	'dynamic-data-lists-form-web/admin/templates/rule-builder.es',
 	function(MetalSoyBundle, RuleBuilder) {
 		if (!window.DDLRuleBuilder) {
 			window.DDLRuleBuilder = {};
 		}
 
-		window.DDLRuleBuilder.render = RuleBuilder.default;
+		RuleBuilder.default.forEach(function(item) {
+			window.DDLRuleBuilder[item.key] = item.component;
+		});
 
 		AUI.add('liferay-ddl-form-builder-rule-builder-template');
 	}
