+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/resources/META-INF/resources/admin/templates/rule-builder.soy
@@ -29,7 +29,7 @@
  * @param rules
  * @param strings
  */
-{template .rule_list }
+{template .rule_list}
 	{if length($rules) > 0}
 		<ul class="ddl-form-body-content form-builder-rule-builder-rules-list tabular-list-group">
 		{foreach $rule in $rules}
