--- modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/resources/META-INF/resources/admin/templates/rule_builder.es.js
@@ -1,14 +1,0 @@
-import Component from 'metal-component';
-import Soy from 'metal-soy';
-
-import templates from './rule_builder.soy';
-
-/**
- * RuleBuilder Component
- */
-class RuleBuilder extends Component {}
-
-// Register component
-Soy.register(RuleBuilder, templates, 'render');
-
-export default RuleBuilder;
