+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/resources/META-INF/resources/admin/templates/sidebar.es.js
@@ -1,8 +1,8 @@
 import Component from 'metal-component';
 import Soy from 'metal-soy';
 
+import './field-options-toolbar.soy';
 import templates from './sidebar.soy';
-import toolbarTemplates from './field_options_toolbar.soy';
 
 /**
  * SideBar Component
